Rework Dexter command imports and get valid commands code to be simpler and more static.
